refac: Move Ansible inventory classes to provisioning module #1596
